requirements.txt
# Core API
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# Data validation
pydantic>=2.7.0

# LangChain + LangGraph
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
langgraph==0.2.28

# Vector store
qdrant-client==1.11.0

# Document parsing
docling==2.5.0
pdf2image==1.17.0
Pillow==10.4.0

# Embeddings + reranking
FlagEmbedding==1.2.11
torch>=2.6.0
colpali-engine==0.3.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.35

# Evaluation
ragas==0.1.21

# HTTP
httpx>=0.27.0
requests>=2.32.3

# UI
chainlit>=2.0.0
